+++ driver/src/clients/mod.rs
@@ -1,0 +1,13 @@
+mod relayer_client;
+mod synchronizer_client;
+
+pub use relayer_client::{RELAYER_POLL_INTERVAL_MS, RELAYER_POLL_TIMEOUT_SECS};
+pub use synchronizer_client::{
+    SYNCHRONIZER_POLL_INTERVAL_MS, SYNCHRONIZER_POLL_TIMEOUT_SECS,
+};
+pub(crate) use relayer_client::{HttpRelayerClient, RelayerClient};
+pub(crate) use synchronizer_client::{HttpSynchronizerClient, SynchronizerClient};
+#[cfg(test)]
+pub(crate) use relayer_client::RelayerConfirmation;
+#[cfg(test)]
+pub(crate) use synchronizer_client::{SynchronizerHead, SynchronizerMembership};
