+++ driver/src/clients/relayer_client.rs
@@ -1,0 +1,151 @@
+use std::time::{Duration, Instant};
+
+use anyhow::{Result, anyhow};
+use base64::{Engine, engine::general_purpose::STANDARD};
+use common::blob::MAX_SIMPLE_BLOB_PAYLOAD_BYTES;
+use relayer::api_types::{JobStatus, JobStatusResponse, SubmitProofRequest, SubmitProofResponse};
+
+pub const RELAYER_POLL_TIMEOUT_SECS: u64 = 180;
+pub const RELAYER_POLL_INTERVAL_MS: u64 = 1500;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct RelayerConfirmation {
+    pub job_id: String,
+    pub tx_hash: Option<String>,
+    pub block_number: Option<i64>,
+}
+
+pub trait RelayerClient: Send + Sync {
+    fn submit_proof(
+        &self,
+        relayer_api_url: &str,
+        payload_bytes: &[u8],
+        client_ref: Option<String>,
+    ) -> Result<SubmitProofResponse>;
+    fn wait_for_confirmation(
+        &self,
+        relayer_api_url: &str,
+        job_id: &str,
+        timeout_secs: u64,
+        poll_interval_ms: u64,
+    ) -> Result<RelayerConfirmation>;
+}
+
+#[derive(Debug, Default)]
+pub struct HttpRelayerClient;
+
+impl RelayerClient for HttpRelayerClient {
+    fn submit_proof(
+        &self,
+        relayer_api_url: &str,
+        payload_bytes: &[u8],
+        client_ref: Option<String>,
+    ) -> Result<SubmitProofResponse> {
+        if payload_bytes.len() > MAX_SIMPLE_BLOB_PAYLOAD_BYTES {
+            return Err(anyhow!(
+                "payload exceeds single-blob limit: {} > {}",
+                payload_bytes.len(),
+                MAX_SIMPLE_BLOB_PAYLOAD_BYTES
+            ));
+        }
+
+        let endpoint = format!("{}/api/v1/proofs", relayer_api_url.trim_end_matches('/'));
+        let request = SubmitProofRequest {
+            payload_base64: STANDARD.encode(payload_bytes),
+            client_ref,
+        };
+
+        let client = reqwest::blocking::Client::new();
+        let response = client
+            .post(&endpoint)
+            .json(&request)
+            .send()
+            .map_err(|err| anyhow!("failed to submit proof to relayer at {endpoint}: {err}"))?;
+
+        let status = response.status();
+        let body = response.text().unwrap_or_default();
+        if !status.is_success() {
+            return Err(anyhow!(
+                "relayer submit failed with {} {}: {}",
+                status.as_u16(),
+                status,
+                body
+            ));
+        }
+
+        serde_json::from_str::<SubmitProofResponse>(&body).map_err(|err| {
+            anyhow!("failed to decode relayer submit response: {err}; body={body}")
+        })
+    }
+
+    fn wait_for_confirmation(
+        &self,
+        relayer_api_url: &str,
+        job_id: &str,
+        timeout_secs: u64,
+        poll_interval_ms: u64,
+    ) -> Result<RelayerConfirmation> {
+        let timeout = Duration::from_secs(timeout_secs);
+        let poll_interval = Duration::from_millis(poll_interval_ms);
+        let start = Instant::now();
+
+        loop {
+            let status = fetch_relayer_job_status(relayer_api_url, job_id)?;
+            match status.status {
+                JobStatus::Confirmed => {
+                    return Ok(RelayerConfirmation {
+                        job_id: status.job_id,
+                        tx_hash: status.tx_hash,
+                        block_number: status.block_number.map(|block_number| block_number as i64),
+                    });
+                }
+                JobStatus::Failed => {
+                    return Err(anyhow!(
+                        "relayer job {} failed: {}",
+                        status.job_id,
+                        status
+                            .last_error
+                            .clone()
+                            .unwrap_or_else(|| "unknown error".to_string())
+                    ));
+                }
+                JobStatus::Queued | JobStatus::Sending | JobStatus::Submitted => {}
+            }
+
+            if start.elapsed() >= timeout {
+                return Err(anyhow!(
+                    "timed out waiting for relayer job {} after {}s",
+                    job_id,
+                    timeout_secs
+                ));
+            }
+            std::thread::sleep(poll_interval);
+        }
+    }
+}
+
+fn fetch_relayer_job_status(relayer_api_url: &str, job_id: &str) -> Result<JobStatusResponse> {
+    let endpoint = format!(
+        "{}/api/v1/proofs/{job_id}",
+        relayer_api_url.trim_end_matches('/')
+    );
+    let client = reqwest::blocking::Client::new();
+    let response = client
+        .get(&endpoint)
+        .send()
+        .map_err(|err| anyhow!("failed to query relayer job at {endpoint}: {err}"))?;
+
+    let status = response.status();
+    let body = response.text().unwrap_or_default();
+    if !status.is_success() {
+        return Err(anyhow!(
+            "relayer status failed with {} {}: {}",
+            status.as_u16(),
+            status,
+            body
+        ));
+    }
+
+    serde_json::from_str::<JobStatusResponse>(&body)
+        .map_err(|err| anyhow!("failed to decode relayer status response: {err}; body={body}"))
+}
